-Updated the HUD to display the RoundTimer without decimals -Added a main menu script -Added a pause menu script

+++ Picnic Panic/Assets/scripts/UI Scripts/HUD.cs
@@ -33,7 +33,7 @@ public class HUD : MonoBehaviour
         {
             displayHealth[i].text = actorHealth[i].Health.ToString();
         }
-        m_displayTimer.text = m_timer.RoundTimer.ToString();
+        m_displayTimer.text = ((int)m_timer.RoundTimer).ToString();
     }
 
     public bool UseLife()
